** Add CompanyEmployee_AddShopDiscount controller and fxml files. *** Update CatalogItemRow.java - new contractor for ItemInShop.java view in table. **** Update ItemInShop.java - changed discountPrice get method to return float.

+++ Team6.Common/src/newEntities/ItemInShop.java
@@ -60,7 +60,7 @@ public class ItemInShop implements IEntity
 	/**
 	 * @return the discountedPrice
 	 */
-	public double getDiscountedPrice()
+	public float getDiscountedPrice()
 	{
 		return m_discountedPrice;
 	}
